Form resets and Logo in Header, Cancel order

+++ OrderService/src/main/java/com/tailor/OrderService/controller/OrderController.java
@@ -76,7 +76,7 @@ public class OrderController {
 
     // Delete order
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable("id") int orderId) {
+    public ResponseEntity<Void> deleteOrder(@PathVariable("id") Long orderId) {
         orderService.deleteOrder(orderId);
         return ResponseEntity.noContent().build();
     }
